Let udev handle only changed devices in kernel_finish. bnc#774301
Added --action=change option to udevadm trigger. Without that
all devices are handled as NEW and it has some side effects.
At least one of them is that device emulated on top of lcs on
s390 is lost. In such case installation cannot continue to
second stage when installing over network.