Call websockify without full dependencies
Also fix \
shell syntax.

The /usr/bin/websockify script uses python3-setuptools to check that
python3-numpy is present. But in fact that dependency is *optional*,
it is 2MB compressed, so we avoid it.

Also, composing the filesystem from various filesystem images and
symlinks, like the inst-sys does, confuses Python initialization quite a
bit, so thus the need for the seemingly superfluous PYTHONPATH.

Even with that PYTHONPATH, numpy is not found even if present, which is
another reason to avoid requiring it.