Respect Textmode=1 when installing over SSH with a DISPLAY (bsc#1047470)
https://bugzilla.suse.com/show_bug.cgi?id=1047470

The fix is: in `check_network`, which is called for SSH installations,
only check `$DISPLAY` when `Textmode` is not set.

Additional small fix: no need to check the qt plugin if we haven't
selected it in the first place. Note that some logical combinations
don't assign `Y2_MODE`. That is fine because for SSH, it is preset to
ncurses before calling this function.