Fix file_owner: move autoinst_files_finish after users_finish (bsc#1196595)
In AutoYaST (AY), the problem was in the application order of

1. <files> <file> with <file_owner>
https://documentation.suse.com/sles/15-SP3/single-html/SLES-autoyast/#createprofile-completeconf

2. https://documentation.suse.com/sles/15-SP3/single-html/SLES-autoyast/#Configuration-Security-users-and-groups

where file_owner refers to a user that was created too late to have an
effect.

The "autoinst_files" step was moved relatively recently (2020-07) from the
second stage of the installation, so I feel moving it after "users" is
less risky than the other way around, moving users before autoinst_files.

- https://github.com/yast/yast-installation/pull/872
- https://github.com/yast/yast-autoinstallation/pull/651